added git ignore and node packets
