--- data/tempPro.json
@@ -1,15 +1,0 @@
-[
-  {
-    "name":"Project Whatever",
-    "image":   "images/temp.png",
-    "git":   "https://github.com/ixnp",
-    "about":    "<p>Interaction design long tail infographic MVP return on investment low hanging fruit crowdfunding alpha focus. Marketing partnership stealth pivot first mover advantage alpha product management technology churn rate conversion. Metrics direct mailing series A financing business plan. Release responsive web design niche market startup bootstrapping. Product management infrastructure startup partner network android conversion crowdfunding growth hacking. Entrepreneur metrics hypotheses equity backing. Release analytics branding scrum project business plan mass market. Scrum project market network effects responsive web design paradigm shift vesting period incubator success ownership venture pitch. Facebook A/B testing iPad client gamification. User experience success ecosystem business model canvas ramen venture return on investment iteration release infographic equity accelerator.</p>"
-
-  },
-  {
-    "name":"Next Big Thing lol",
-    "image":"images/temp2.png",
-    "git":"https://github.com/ixnp",
-    "about":"<p>Interaction design long tail infographic MVP return on investment low hanging fruit crowdfunding alpha focus. Marketing partnership stealth pivot first mover advantage alpha product management technology churn rate conversion. Metrics direct mailing series A financing business plan. Release responsive web design niche market startup bootstrapping. Product management infrastructure startup partner network android conversion crowdfunding growth hacking. Entrepreneur metrics hypotheses equity backing. Release analytics branding scrum project business plan mass market. Scrum project market network effects responsive web design paradigm shift vesting period incubator success ownership venture pitch. Facebook A/B testing iPad client gamification. User experience success ecosystem business model canvas ramen venture return on investment iteration release infographic equity accelerator.</p>"
-  }
-];
