+++ server.js
@@ -1,0 +1,20 @@
+'use strict';
+
+const pg = require('pg');
+const fs = require('fs');
+const express = require('express');
+const bodyParser = require('body-parser');
+const PORT = process.env.PORT || 3000;
+const app = express();
+// const conString = 'postgres://USERNAME:PASSWORD@HOST:PORT';
+const conString = 'postgres://postgres:1234@localhost:5432/lab12'; // TODO: Don't forget to set your own conString
+const client = new pg.Client(conString);
+client.connect();
+client.on('error', err => console.error(err));
+
+app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({extended: true}));
+app.use(express.static('./public'));
+
+
+app.listen(PORT, () => console.log(`Server started on port ${PORT}!`));
