--- style/base.css
@@ -1,13 +1,0 @@
-/*font-family: 'VT323', monospace;
-font-family: 'Arvo', serif;*/
-*{
-  font-family: 'VT323', monospace;
-  color:white;
-  display: flex;
-  flex-direction: column;
-  /*font-family: 'Arvo', serif;*/
-  background-color: black;
-}
-#template{
-display: none !important;
-}
