--- style/layout.css
@@ -1,22 +1,0 @@
-.top{
-  display:flex;
-  flex-direction: column;
-  flex-wrap: wrap;
-  justify-content: flex-start;
-  align-content: flex-start;
-  align-items: center;
-
-}
-.bannerImage{
-width: 100%;
-height: auto;
-}
-h1,.name{
-/*going to need to change this margin so it looks better across devices*/
-margin-top: 48%;
-font-size: 8.5vw;
-display: flex;
-background-color: transparent;
-position: absolute;
-z-index: 10;
-}
